Increased OOPness of code, refactored out UI and some parts of command execution
